+++ dear-imgui-sys/src/platform_io_hooks.cpp
@@ -1,0 +1,126 @@
+#include "../third-party/cimgui/imgui/imgui.h"
+
+#ifdef IMGUI_HAS_DOCK
+
+struct DearImguiRsPlatformIoHookStorage
+{
+    ImGuiPlatformIO* PlatformIO;
+    void (*Platform_GetWindowPos)(ImGuiViewport* vp, ImVec2* out_pos);
+    void (*Platform_GetWindowSize)(ImGuiViewport* vp, ImVec2* out_size);
+};
+
+static ImVector<DearImguiRsPlatformIoHookStorage> G_DearImguiRsPlatformIoHookStorage;
+
+static DearImguiRsPlatformIoHookStorage* DearImguiRsFindPlatformIoHookStorage(ImGuiPlatformIO* platform_io)
+{
+    if (platform_io == nullptr)
+        return nullptr;
+    for (int n = 0; n < G_DearImguiRsPlatformIoHookStorage.Size; n++)
+        if (G_DearImguiRsPlatformIoHookStorage[n].PlatformIO == platform_io)
+            return &G_DearImguiRsPlatformIoHookStorage[n];
+    return nullptr;
+}
+
+static DearImguiRsPlatformIoHookStorage& DearImguiRsGetPlatformIoHookStorage(ImGuiPlatformIO* platform_io)
+{
+    if (DearImguiRsPlatformIoHookStorage* storage = DearImguiRsFindPlatformIoHookStorage(platform_io))
+        return *storage;
+
+    DearImguiRsPlatformIoHookStorage storage = {};
+    storage.PlatformIO = platform_io;
+    G_DearImguiRsPlatformIoHookStorage.push_back(storage);
+    return G_DearImguiRsPlatformIoHookStorage[G_DearImguiRsPlatformIoHookStorage.Size - 1];
+}
+
+static DearImguiRsPlatformIoHookStorage* DearImguiRsGetCurrentPlatformIoHookStorage()
+{
+    if (ImGui::GetCurrentContext() == nullptr)
+        return nullptr;
+    return DearImguiRsFindPlatformIoHookStorage(&ImGui::GetPlatformIO());
+}
+
+static void DearImguiRsPrunePlatformIoHookStorageIfEmpty(ImGuiPlatformIO* platform_io)
+{
+    DearImguiRsPlatformIoHookStorage* storage = DearImguiRsFindPlatformIoHookStorage(platform_io);
+    if (storage == nullptr)
+        return;
+    if (storage->Platform_GetWindowPos != nullptr || storage->Platform_GetWindowSize != nullptr)
+        return;
+    G_DearImguiRsPlatformIoHookStorage.erase(storage);
+}
+
+static ImVec2 DearImguiRsPlatformGetWindowPosHook(ImGuiViewport* vp)
+{
+    ImVec2 pos(0.0f, 0.0f);
+    if (DearImguiRsPlatformIoHookStorage* storage = DearImguiRsGetCurrentPlatformIoHookStorage())
+        if (storage->Platform_GetWindowPos != nullptr)
+            storage->Platform_GetWindowPos(vp, &pos);
+    return pos;
+}
+
+static ImVec2 DearImguiRsPlatformGetWindowSizeHook(ImGuiViewport* vp)
+{
+    ImVec2 size(0.0f, 0.0f);
+    if (DearImguiRsPlatformIoHookStorage* storage = DearImguiRsGetCurrentPlatformIoHookStorage())
+        if (storage->Platform_GetWindowSize != nullptr)
+            storage->Platform_GetWindowSize(vp, &size);
+    return size;
+}
+
+extern "C" void dear_imgui_rs_platform_io_set_platform_get_window_pos(
+    ImGuiPlatformIO* platform_io,
+    void (*user_callback)(ImGuiViewport* vp, ImVec2* out_pos))
+{
+    if (platform_io == nullptr)
+        return;
+
+    if (user_callback == nullptr)
+    {
+        if (DearImguiRsPlatformIoHookStorage* storage = DearImguiRsFindPlatformIoHookStorage(platform_io))
+            storage->Platform_GetWindowPos = nullptr;
+        platform_io->Platform_GetWindowPos = nullptr;
+        DearImguiRsPrunePlatformIoHookStorageIfEmpty(platform_io);
+        return;
+    }
+
+    DearImguiRsPlatformIoHookStorage& storage = DearImguiRsGetPlatformIoHookStorage(platform_io);
+    storage.Platform_GetWindowPos = user_callback;
+    platform_io->Platform_GetWindowPos = DearImguiRsPlatformGetWindowPosHook;
+}
+
+extern "C" void dear_imgui_rs_platform_io_set_platform_get_window_size(
+    ImGuiPlatformIO* platform_io,
+    void (*user_callback)(ImGuiViewport* vp, ImVec2* out_size))
+{
+    if (platform_io == nullptr)
+        return;
+
+    if (user_callback == nullptr)
+    {
+        if (DearImguiRsPlatformIoHookStorage* storage = DearImguiRsFindPlatformIoHookStorage(platform_io))
+            storage->Platform_GetWindowSize = nullptr;
+        platform_io->Platform_GetWindowSize = nullptr;
+        DearImguiRsPrunePlatformIoHookStorageIfEmpty(platform_io);
+        return;
+    }
+
+    DearImguiRsPlatformIoHookStorage& storage = DearImguiRsGetPlatformIoHookStorage(platform_io);
+    storage.Platform_GetWindowSize = user_callback;
+    platform_io->Platform_GetWindowSize = DearImguiRsPlatformGetWindowSizeHook;
+}
+
+#else
+
+extern "C" void dear_imgui_rs_platform_io_set_platform_get_window_pos(
+    ImGuiPlatformIO*,
+    void (*)(ImGuiViewport*, ImVec2*))
+{
+}
+
+extern "C" void dear_imgui_rs_platform_io_set_platform_get_window_size(
+    ImGuiPlatformIO*,
+    void (*)(ImGuiViewport*, ImVec2*))
+{
+}
+
+#endif